Fjern felles finnBegrunnelse som blir overflødig når kode er splittet opp
